// src/dorc-web/src/apis/dorc-api/models/AnalyticsDeploymentsPerProjectApiModel.ts
// tslint:disable
/**
 * Dorc.Api
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

/**
 * @export
 * @interface AnalyticsDeploymentsPerProjectApiModel
 */
export interface AnalyticsDeploymentsPerProjectApiModel {
  /**
   * @type {number}
   * @memberof AnalyticsDeploymentsPerProjectApiModel
   */
  Year?: number;
  /**
   * @type {number}
   * @memberof AnalyticsDeploymentsPerProjectApiModel
   */
  Month?: number;
  /**
   * @type {number}
   * @memberof AnalyticsDeploymentsPerProjectApiModel
   */
  Day?: number;
  /**
   * @type {string}
   * @memberof AnalyticsDeploymentsPerProjectApiModel
   */
  ProjectName?: string | null;
  /**
   * @type {number}
   * @memberof AnalyticsDeploymentsPerProjectApiModel
   */
  CountOfDeployments?: number;
  /**
   * @type {number}
   * @memberof AnalyticsDeploymentsPerProjectApiModel
   */
  Failed?: number;
}